move files around, more logic
moved xp stuff to a new file cuz it was getting long
better filtering logic for reaction and posting xp
adjusted some of the calculations for the multipliers in reactions. They should give a bit more now.
used the mod summary embed for the trust notification
reworked poll reactions to include a few more parameters
add chat activity to the mod embeds
feathers initial commit with lure integration
