+++ config/snowflakes-testing.json
@@ -36,7 +36,8 @@
     "gb": "1215052043354640484",
     "noice": "1207105248838291566",
     "steam": "1219786130317381632",
-    "slothmare": "1289379209524809728"
+    "slothmare": "1289379209524809728",
+    "xpFeather": "1322322151935709227"
   },
   "ldsg": "1207041599608061962",
   "other": {
