+++ config/snowflakes.json
@@ -36,7 +36,8 @@
     "gb": "493084576470663180",
     "noice": "1235026942227513354",
     "steam": "813426118518243388",
-    "slothmare": "726133915655602276"
+    "slothmare": "726133915655602276",
+    "xpFeather": "1322321879553413231"
   },
   "ldsg": "96335850576556032",
   "other": {
