+++ utils/modCommon.js
@@ -311,7 +311,7 @@ const modCommon = {
     if (getInfractions) {
       embed
         .setDescription(text)
-        .addFields({ name: "Activity", value: `Active Minutes: ${userDoc?.posts ?? "Unknown"}`, inline: true });
+        .addFields({ name: "Activity", value: `Chat: ${userDoc?.posts ?? "Unknown"} minutes\nVoice: ${userDoc?.voice ?? "Unknown"}`, inline: true });
     }
     return embed;
